Move NTAuthentication/NegotiateStreamPal back to the mothership
