+++ src/libraries/System.Net.Security/src/System/Net/ContextFlagsAdapterPal.Windows.cs
