+++ src/libraries/System.Net.Security/src/System/Net/ContextFlagsPal.cs
