+++ src/libraries/System.Net.Security/src/System/Net/NTAuthentication.Managed.cs
