+++ src/libraries/System.Net.Security/src/System/Net/SecurityStatusAdapterPal.Windows.cs
