+++ src/libraries/System.Net.Security/src/System/Net/SecurityStatusPal.cs
